ask-27 : Ajouter la gestion des erreurs dans chaque MS

+++ src/main/java/com/bookNDrive/formula_service/services/CrudService.java
@@ -1,6 +1,7 @@
 package com.bookNDrive.formula_service.services;
 
 import com.bookNDrive.formula_service.dtos.sended.FormulaDto;
+import com.bookNDrive.formula_service.exceptions.FormulaNotFoundException;
 import com.bookNDrive.formula_service.mappers.FormulaMapper;
 import com.bookNDrive.formula_service.repositories.FormulaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +23,8 @@ public class CrudService {
     }
 
     public FormulaDto getFormula(Long id) {
-        var formula = formulaRepository.findById(id).orElseThrow(() -> new RuntimeException("formula doesn't find in bdd"));
+        var formula = formulaRepository.findById(id)
+                .orElseThrow(() -> new FormulaNotFoundException(id));
         return formulaMapper.formulaToFormulaDto(formula);
     }
 
@@ -33,25 +35,24 @@ public class CrudService {
 
     @PreAuthorize("hasRole('ADMIN')")
     public FormulaDto createFormula(FormulaDto formula) {
-
         var formulaEntity = formulaRepository.save(formulaMapper.formulaDtoToFormula(formula));
         return formulaMapper.formulaToFormulaDto(formulaEntity);
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     public void deleteFormula(Long formulaId) {
+        if (!formulaRepository.existsById(formulaId)) {
+            throw new FormulaNotFoundException(formulaId);
+        }
         formulaRepository.deleteById(formulaId);
     }
 
     @PreAuthorize("hasRole('ADMIN')")
     public FormulaDto updateFormula(Long formulaId, FormulaDto formulaDto) {
-        formulaRepository.findById(formulaId).orElseThrow(() -> {
-            throw new RuntimeException("Formule non trouvée en bdd");
-        });
+        formulaRepository.findById(formulaId)
+                .orElseThrow(() -> new FormulaNotFoundException(formulaId));
 
         var formulaEntity = formulaRepository.save(formulaMapper.formulaDtoToExistingFormula(formulaDto));
         return formulaMapper.formulaToFormulaDto(formulaEntity);
-
     }
-
 }
